Fixed code in Example 3-12
The code as it is renders the leftmost ellipse with a strokeWeight of 20, as draw() is continuously called and without an exit(). While exit() is one solution, the other, given that readers haven't been introduced to the method by this stage, would be to set the strokeWeight(1) for the first ellipse.